Fix build of Preview
