+++ Shared/Package.swift
@@ -8,7 +8,6 @@ let package = Package(
     defaultLocalization: "en",
     platforms: [.iOS(.v13), .macOS(.v10_15)],
     products: [
-        // Products define the executables and libraries a package produces, and make them visible to other packages.
         .library(
             name: "Document",
             targets: ["Document"]),
@@ -26,8 +25,7 @@ let package = Package(
         
     ],
     dependencies: [
-        // Dependencies declare other packages that this package depends on.
-        // .package(url: /* package url */, from: "1.0.0"),
+        
     ],
     targets: [
         .target(
