+++ Shared/Sources/Canvas/CanvasPresenter.swift
@@ -10,41 +10,6 @@ import CoreText
 import Combine
 import TextRecognition
 
-public class DocumentPresenter {
-    
-    public init(document: VODesignDocumentProtocol) {
-        self.document = document
-    }
-    
-    public private(set) var document: VODesignDocumentProtocol
-    
-    var drawingController: DrawingController!
-    public weak var ui: DrawingView!
-    
-    public func save() {
-        let descriptions = ui.drawnControls.compactMap { control in
-            control.a11yDescription
-        }
-        
-        document.controls = descriptions
-    }
-    
-    public let selectedPublisher = OptionalDescriptionSubject(nil)
-    public let recognitionPublisher = TextRecognitionSubject(nil)
-    
-    public func update(image: Image) {
-        document.image = image
-    }
-    
-    func update(controls: [A11yDescription]) {
-        document.controls = controls
-    }
-    
-    func publish(textRecognition: RecognitionResult) {
-        recognitionPublisher.send(textRecognition)
-    }
-}
-
 public protocol CanvasPresenterUIProtocol: AnyObject {
     func image(at frame: CGRect) async -> CGImage?
 }
