+++ Shared/Sources/Document/Documents/VODesignDocument+AppKit.swift
@@ -4,16 +4,7 @@ import AppKit
 public typealias Document = NSDocument
 
 import os
-import Foundation
 import Combine
-
-public protocol VODesignDocumentProtocol {
-    var controls: [A11yDescription] { get set }
-    var undoManager: UndoManager? { get }
-    var image: Image? { get set }
-    
-    var controlsPublisher: PassthroughSubject<[A11yDescription], Never> { get }
-}
 
 public class VODesignDocument: Document, VODesignDocumentProtocol {
     public static var vodesign = "vodesign"
@@ -144,4 +135,3 @@ public class VODesignDocument: Document, VODesignDocumentProtocol {
     
 }
 #endif
-
