+++ Shared/Sources/Document/Models/ElementColors.swift
