+++ VoiceOver Preview/PreviewFeatures/Package.swift
@@ -7,7 +7,6 @@ let package = Package(
     name: "PreviewFeatures",
     platforms: [.iOS(.v13)],
     products: [
-        // Products define the executables and libraries a package produces, and make them visible to other packages.
         .library(
             name: "DesignPreview",
             targets: ["DesignPreview"]),
@@ -16,12 +15,11 @@ let package = Package(
         .package(name: "Shared", path: "./../../Shared")
     ],
     targets: [
-        // Targets are the basic building blocks of a package. A target can define a module or a test suite.
-        // Targets can depend on other targets in this package, and on products in packages this package depends on.
         .target(
             name: "DesignPreview",
             dependencies: [
                 .product(name: "Document", package: "Shared"),
+                .product(name: "Canvas", package: "Shared"),
             ]),
     ]
 )
